handle alt_measures in food_details_page.dart

+++ lib/app/models/alt_measure_model/alt_measure_model.dart
@@ -10,8 +10,8 @@ class AltMeasureModel with _$AltMeasureModel {
   factory AltMeasureModel({
     @JsonKey(name: 'serving_weight') double? servingWeight,
     String? measure,
-    int? seq,
-    int? qty,
+    double? seq,
+    double? qty,
   }) = _AltMeasureModel;
 
   factory AltMeasureModel.fromJson(Map<String, dynamic> json) => _$AltMeasureModelFromJson(json);
